Change cam_periph_ioctl() to take 'cmd' and a u_long instead of an int.  All of its callers pass in cmd as a u_long, so this has always been a dangerous type demotion.  It was spooted by clang/llvm trying to do a type promotion and sign extension within cam_periph_ioctl.
Submitted by:	rdivacky

+++ sys/cam/cam_periph.c
@@ -797,7 +797,7 @@ cam_periph_ccbwait(union ccb *ccb)
 }
 
 int
-cam_periph_ioctl(struct cam_periph *periph, int cmd, caddr_t addr,
+cam_periph_ioctl(struct cam_periph *periph, u_long cmd, caddr_t addr,
 		 int (*error_routine)(union ccb *ccb, 
 				      cam_flags camflags,
 				      u_int32_t sense_flags))
